==> 03 Final Scripts/01 Orders Data Cleaning.ipynb <==
# Orders Data Cleaning

## Contents

### 01 Importing Libraries

### 02 Importing Data

### 03 Checking Data

### 04 Wrangling Data

#### 01 Columns

#### 02 Data Types

### 05 Checking Data Consistency

#### 01 Missing Values

#### 02 Duplicate Values

### 06 Exporting Dataframe

## 01 Importing Libraries

#importing libraries
import pandas as pd
import numpy as np
import os

## 02 Importing Data

#creating path
path = r'C:\\Users\\samac\\Instacart Basket Analysis'

#importing orders.csv
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

## 03 Checking Data

#checking df_ords shape
df_ords.shape

#checking df_ords records
df_ords.head()

## 04 Wrangling Data

### 01 Columns

#dropping 'eval_set' column from df_ords
df_ords = df_ords.drop(columns = ['eval_set'])

#renaming 'order_dow' to 'orders_day_of_week'
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

#checking column titles
df_ords.columns

### 02 Data Types

#checking data types in df_ords
df_ords.dtypes

#changing 'order_id' from int to string
df_ords['order_id'] = df_ords['order_id'].astype('str')

#changing 'user_id' from int to string
df_ords['user_id'] = df_ords['user_id'].astype('str')

#rechecking data types in df_ords
df_ords.dtypes

## 05 Data Consistency Checks

### 01 Missing Values

#searching for missing values in df_ords
df_ords.isnull().sum()

#creating flag based on missing values in 'days_since_prior_order' column
df_ords['days_since_prior_order_flag'] = df_ords['days_since_prior_order'].isna().astype(int)

#checking df_ords
df_ords.head()

### 02 Duplicate Values

#searching for duplicate values in df_ords
df_ords_dups = df_ords[df_ords.duplicated()]

df_ords_dups

No Duplicates Found

## 06 Exporting Dataframe

#exporting df_ords as final_orders_wrangled.pkl
df_ords.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'final_orders_wrangled.pkl'))
