==> 03 Final Scripts/04 Products Data Cleaning.ipynb <==
# Products Data Cleaning

## Contents

### 01 Importing Libraries

### 02 Importing Data

### 03 Checking Data

### 04 Wrangling Data

#### 01 Data Types

### 05 Checking Data Consistency

#### 01 Missing Values

#### 02 Duplicate Values

### 06 Exporting Dataframe

## 01 Importing Libraries

#importing libraries
import pandas as pd
import numpy as np
import os

## 02 Importing Data

#creating path
path = r'C:\\Users\\samac\\Instacart Basket Analysis'

#importing orders.csv
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

## 03 Checking Data

#checking df_ords shape
df_prods.shape

#checking df_ords records
df_prods.head()

## 04 Wrangling Data

### 01 Data Types

#checking data types in df_prods
df_prods.dtypes

#changing 'product_id' from int to string
df_prods['product_id'] = df_prods['product_id'].astype('str')

#changing 'aisle_id' from int to string
df_prods['aisle_id'] = df_prods['aisle_id'].astype('str')

#changing 'department_id' from int to string
df_prods['department_id'] = df_prods['department_id'].astype('str')

#rechecking data types in df_prods
df_prods.dtypes

## 05 Data Consistency Checks

### 01 Missing Values

#searching for missing values in df_prods
df_prods.isnull().sum()

# finding null product names
df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

#creating new df without products with missing names

df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

#checking shape of df_prods_clean
df_prods_clean.shape

### 02 Duplicate Values

#searching for duplicate values in df_prods_clean
df_prods_dups = df_prods_clean[df_prods_clean.duplicated()]

df_prods_dups

#dropping duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

#checking shape of df_prods_clean_no_dups
df_prods_clean_no_dups.shape

## 06 Exporting Dataframe

#exporting df_prods_clean_no_dups as final_products_wrangled.pkl
df_prods_clean_no_dups.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'final_products_wrangled.pkl'))
