==> 03 Final Scripts/05 Merging Products and Final Orders Products Combined.ipynb <==
# Merging Orders and Orders Prior

## Contents

### 01 Importing Libraries

### 02 Importing Data

### 03 Merging Data

### 04 Checking New Dataframe

### 05 Exporting Dataframe

## 01 Importing Libraries

#importing libraries
import pandas as pd
import numpy as np
import os

## 02 Importing Data

#creating path
path = r'C:\\Users\\samac\\Instacart Basket Analysis'

#importing final_prods_wrangled
df_prods = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'final_products_wrangled.pkl'))

#importing final_orders_products_combined
df_ords_prods_combined = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'final_orders_products_combined.pkl'))

#checking df_prods
df_prods.head()

#checking df_ords_prods_combined
df_ords_prods_combined.head()

## 03 Merging Data

#dropping '_merge' column from df_ords_prods_combined
df_ords_prods_combined = df_ords_prods_combined.drop(columns = ['_merge'])

#checking df_ords_prods_combined
df_ords_prods_combined.head()

#merging df_prods and df_ords_prods_combined on 'order_id' column
ords_prods_merged = df_prods.merge(df_ords_prods_combined, on = 'product_id', indicator = True)

## 04 Checking New Dataframe

#checking df_merged_large shape
ords_prods_merged.shape

ords_prods_merged.head(20)

## 05 Exporting Dataframe

#exporting ords_prods_merged as 'final_orders_products_merged.pkl' pickle file

ords_prods_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'final_orders_products_merged.pkl'))
