==> 03 Final Scripts/06 Customers Data Cleaning.ipynb <==
# Orders Data Cleaning

## Contents

### 01 Importing Libraries

### 02 Importing Data

### 03 Checking Data

### 04 Wrangling Data

#### 01 Columns

#### 02 Data Types

### 05 Checking Data Consistency

#### 01 Missing Values

#### 02 Duplicate Values

### 06 Exporting Dataframe

## 01 Importing Libraries

#importing libraries
import pandas as pd
import numpy as np
import os

## 02 Importing Data

#creating path
path = r'C:\\Users\\samac\\Instacart Basket Analysis'

#importing customers.csv
df_customers = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'), index_col = False)

## 03 Checking Data

#getting shape of df_customers
df_customers.shape

#describing df_customers
df_customers.describe()

#checking df_customers
df_customers.head()

## 04 Wrangling Data

### 01 Columns

#changing First Name to first_name
df_customers.rename(columns = {'First Name' : 'first_name'}, inplace = True)

#changing Surnam to last_name
df_customers.rename(columns = {'Surnam' : 'last_name'}, inplace = True)

#changing Gender to gender
df_customers.rename(columns = {'Gender' : 'gender'}, inplace = True)

#changing STATE to state
df_customers.rename(columns = {'STATE' : 'state'}, inplace = True)

#changing Age to age
df_customers.rename(columns = {'Age' : 'age'}, inplace = True)

#checking column names
df_customers.columns

### 02 Data Types

#checking data types in df_ords
df_customers.dtypes

# changing user_id column to string for df_customers
df_customers['user_id'] = df_customers['user_id'].astype('str')

#rechecking data types in df_ords
df_customers.dtypes

## 05 Data Consistency Checks

### 01 Missing Values

#searching for missing values in df_customers
df_customers.isnull().sum()

### 02 Duplicate Values

#searching for duplicate values in df_customers
df_customers_dups = df_customers[df_customers.duplicated()]

df_customers_dups

No Duplicates Found

## 06 Exporting Dataframe

#exporting df_customers as final_customers_wrangled.pkl
df_customers.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'final_customers_wrangled.pkl'))
