==> 03 Final Scripts/07 Merging Customers and Final Orders Products Merged.ipynb <==
# Merging Orders and Orders Prior

## Contents

### 01 Importing Libraries

### 02 Importing Data

### 03 Merging Data

### 04 Checking New Dataframe

### 05 Exporting Dataframe

## 01 Importing Libraries

#importing libraries
import pandas as pd
import numpy as np
import os

## 02 Importing Data

#creating path
path = r'C:\\Users\\samac\\Instacart Basket Analysis'

#importing final_orders_products_merged
ords_prods_merged = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'final_orders_products_merged.pkl'))

#importing customer
df_customers = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'final_customers_wrangled.pkl'))

#checking ords_prods_merged
ords_prods_merged.head()

#checking df_customers
df_customers.head()

## 03 Merging Data

#dropping '_merge' column from df_ords_prods_merged
ords_prods_merged = ords_prods_merged.drop(columns = ['_merge'])

#checking df_ords_prods_merged
ords_prods_merged.head()

#merging ords_prods_merge and df_customers on user_id
ords_prods_all = ords_prods_merged.merge(df_customers, on = 'user_id', indicator = True)

## 04 Checking New Dataframe

#checking ords_prods_all shape
ords_prods_all.shape

ords_prods_all.head(20)

## 05 Exporting Dataframe

#exporting ords_prods_all as 'final_orders_products_all.pkl' pickle file

ords_prods_all.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'final_orders_products_all.pkl'))
