==> 03 Final Scripts/08 Column Derivations.ipynb <==
# Column Derivations

## Contents

### 01 Importing Libraries

### 02 Importing Data

### 03 Deriving New Columns

### 04 Exporting Dataframe

## 01 Importing Libraries

#importing libraries

import pandas as pd
import numpy as np
import os

## 02 Importing Data

#creating path
path = r'C:\\Users\\samac\\Instacart Basket Analysis'

#importing final_ords_prods_all
final_ords_prods_all = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'final_orders_products_all.pkl'))

final_ords_prods_all.columns

#checking for nulls
final_ords_prods_all.isnull().sum()

## 03 Deriving New Columns

The following column derivations have been created at the request of the client

### 01 Price Range

The price_range_loc column will show what price range each product falls into. Products priced 5 dollars and lower are "Low-range products". Products priced in between 5 and 15 dollars are "Mid-range products". Products priced over 15 dollars are "High-range" products.

#creating if statements with the loc() function
final_ords_prods_all.loc[final_ords_prods_all['prices'] > 15, 'price_range_loc'] = 'High-range product'
final_ords_prods_all.loc[(final_ords_prods_all['prices'] <= 15) & (final_ords_prods_all['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 
final_ords_prods_all.loc[final_ords_prods_all['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

#checking value counts
final_ords_prods_all['price_range_loc'].value_counts(dropna = False)

### 02 Busiest Day

The busiest_day column will label the day that has the most orders as the "Busiest day', the day with the fewest orders as the 'Least busy", and all other days as 'Regularly busy'

#checking value count of orders_day_of_week
final_ords_prods_all['orders_day_of_week'].value_counts(dropna = False)

#creating busy day list
result = []

for value in final_ords_prods_all["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

#adding busiest_day column
final_ords_prods_all['busiest_day'] = result

#checking value count of busiest day
final_ords_prods_all['busiest_day'].value_counts(dropna = False)

### 03 Busiest Days

The busiest_days column will label the 2 days that have the most orders as the 'Busiest days', the 2 days with the fewest orders as the 'Least busiest days", and all other days as 'Regularly busy'

#creating busiest_days list
busiest_results = []

for value in final_ords_prods_all["orders_day_of_week"]:
  if value == 0 or value == 1:
    busiest_results.append("Busiest days")
  elif value == 4 or value == 3:
    busiest_results.append("Least busiest days")
  else:
    busiest_results.append("Regularly busy")

#adding busiest_day column
final_ords_prods_all['busiest_days'] = busiest_results

#checking value count of busiest day
final_ords_prods_all['busiest_days'].value_counts(dropna = False)

### 04 Busiest Period of Day

The busiest_period_of_day will label the 3 hours that have the most orders as 'Most orders', the 3 hours with the fewest orders as 'Fewest orders', and all other hours as 'Average orders'

#checking value count of order_hour_of_day
final_ords_prods_all['order_hour_of_day'].value_counts()

#dividing each hour into one of 3 categories
#The top 3 most frequent hours will be labeled 'Most orders', the bottom 3 'Fewest orders', and the remaining 'Average orders'

hours_results = []

for value in final_ords_prods_all['order_hour_of_day']:
    if value in [10, 11, 14]:
        hours_results.append('Most orders')
    elif value in [3, 4, 2]:
        hours_results.append('Fewest orders')
    else:
        hours_results.append('Average orders')

#adding busiest_period_of_day column
final_ords_prods_all['busiest_period_of_day'] = hours_results

#checking value count of busiest_period_of_day
final_ords_prods_all['busiest_period_of_day'].value_counts()

### 05 Max Order

The max_order column will return each user's max order number

#creating max order column
final_ords_prods_all['max_order'] = final_ords_prods_all.groupby(['user_id'])['order_number'].transform(np.max)

#checking to see if max row is a new column
final_ords_prods_all.head()

### 06 Loyalty Flag

The loyalty_flag column divides the max_order column into three categories. If the value is greater than 40, the user is labeled a 'Loyal customer'. If the value is lower than or equal to 40 and greater than 10, the user is labeled a 'Regular customer'. If the value is lower than 10, the user is labeled a 'New customer'

#creating loyalty column 
final_ords_prods_all.loc[final_ords_prods_all['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'
final_ords_prods_all.loc[(final_ords_prods_all['max_order'] <= 40) & (final_ords_prods_all['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'
final_ords_prods_all.loc[final_ords_prods_all['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

#checking value counts
final_ords_prods_all['loyalty_flag'].value_counts(dropna = False)

### 07 Average Price

The avg_price column will return the average price of each user's total orders

#creating avg_price column
final_ords_prods_all['avg_price'] = final_ords_prods_all.groupby(['user_id'])['prices'].transform(np.mean)

#checking to see if avg_price is a new column
final_ords_prods_all.head()

### 08 Spending Flag

The spending_flag column divides the avg_price column into two categories. If the value is greater than or equal to 10, the user is labeled a 'High spender'. If the value is lower than 10, the user is labeled a 'Low spender'

#creating spending flag for low spenders
final_ords_prods_all.loc[final_ords_prods_all['avg_price'] < 10, 'spending_flag'] = 'Low Spender'

#creating spending flag for high spenders
final_ords_prods_all.loc[final_ords_prods_all['avg_price'] >= 10, 'spending_flag'] = 'High Spender'

#checking value counts of spenders
final_ords_prods_all['spending_flag'].value_counts(dropna = False)

### 09 Days Since Prior Order Median

The days_since_prior_order_median flag will the median days_since_prior_order for each user

#creating days_since_prior_order_median column
final_ords_prods_all['days_since_prior_order_median'] = final_ords_prods_all.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

#checking to see if days_since_prior_order_median is a new column
final_ords_prods_all.head()

### 10 Frequency Flag

The frequency_flag will label divides the median days_since_prior_order_median_column into three categories. If the value is greater than 20, the user is labeled a 'Non-frequest customer'. If the value is lower than or equal to 20 and greater than 10, the user is labeled a 'Regular customer'. If the value is lower than 10, the user is labeled a 'Frequent customer'

#creating Non-Frequent customer frequency_flag
final_ords_prods_all.loc[final_ords_prods_all['days_since_prior_order_median'] > 20, 'frequency_flag'] = 'Non-frequent customer'

#creating regular customer frequency_flag
final_ords_prods_all.loc[(final_ords_prods_all['days_since_prior_order_median'] > 10) & (final_ords_prods_all['days_since_prior_order_median'] <= 20), 'frequency_flag'] = 'Regular customer'

#creating frequent customer frequency_flag
final_ords_prods_all.loc[final_ords_prods_all['days_since_prior_order_median'] <= 10, 'frequency_flag'] = 'Frequent customer'

#checking value counts of frequency flag
final_ords_prods_all['frequency_flag'].value_counts(dropna = False)

## 04 Exporting Dataframe

final_ords_prods_all.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'final_orders_products_all_with_column_derivations'))
