==> 03 Final Scripts/10 Departments Data Cleaning.ipynb <==
# Orders Data Cleaning

## Contents

### 01 Importing Libraries

### 02 Importing Data

### 03 Checking Data

### 04 Formatting Data

### 05 Exporting Dataframe

## 01 Importing Libraries

#importing libraries
import pandas as pd
import numpy as np
import os

## 02 Importing Data

#creating path
path = r'C:\\Users\\samac\\Instacart Basket Analysis'

#importimg departments.csv
df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

## 03 Checking Data

#checking df_ords shape
df_dep.shape

#checking df_ords records
df_dep.head()

## 04 Formatting Data

#transposing df_dep
df_dep_t = df_dep.T
df_dep_t

#resetting index
df_dep_t.reset_index()

#creating new header using top records
new_header = df_dep_t.iloc[0]
new_header

#removing first record of dataframe
df_dep_t_new = df_dep_t[1:]
df_dep_t_new

#adding column headers
df_dep_t_new.columns = new_header
df_dep_t_new

## 05 Exporting Dataframe

#exporting df_dep_t_new as 
df_dep_.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'final_departments_wrangled.pkl'))
